requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.2.1
tenacity==8.2.3
python-dotenv>=1.0.0
openpyxl>=3.1.2
beautifulsoup4>=4.12.0
playwright>=1.40.0
cloudscraper>=1.2.71